+++ test/socketio/cloudevent_socket.io.test.js
@@ -11,7 +11,7 @@ cds.env.websocket = {
   impl: null,
 };
 
-const cloudEvent1Message = JSON.stringify({
+const cloudEvent = {
   specversion: "1.0",
   type: "com.example.someevent",
   source: "/mycontext",
@@ -26,70 +26,69 @@ const cloudEvent1Message = JSON.stringify({
     appinfoB: 123,
     appinfoC: true,
   },
-});
+};
 
-const cloudEvent2Message = JSON.stringify({
-  specversion: "1.0",
-  type: "com.example.someevent",
-  source: "/mycontext",
-  subject: null,
-  id: "C234-1234-1234",
-  time: "2018-04-05T17:31:00Z",
-  comexampleextension1: "value",
-  comexampleothervalue: 5,
-  datacontenttype: "application/json",
-  data: {
-    appinfoA: "abc",
-    appinfoB: 123,
-    appinfoC: true,
-  },
-});
+const cloudEventModel = {
+  ...cloudEvent,
+  type: "com.example.someevent.model",
+};
 
-const cloudEvent3Message = JSON.stringify({
-  specversion: "1.0",
-  type: "com.example.someevent",
-  source: "/mycontext",
-  subject: null,
-  id: "C234-1234-1234",
-  time: "2018-04-05T17:31:00Z",
-  comexampleextension1: "value",
-  comexampleothervalue: 5,
-  datacontenttype: "application/json",
-  data: {
-    appinfoA: "abc",
-    appinfoB: 123,
-    appinfoC: true,
-  },
-});
+const cloudEventMap = {
+  ...cloudEvent,
+  type: "com.example.someevent.map",
+};
 
-const cloudEvent4Message = JSON.stringify({
+const cloudEvent1 = {
   specversion: "1.0",
-  type: "com.example.someevent",
-  source: "/mycontext",
+  type: "CloudEventService.cloudEvent1",
+  source: "CloudEventService",
   subject: null,
-  id: "C234-1234-1234",
-  time: "2018-04-05T17:31:00Z",
-  comexampleextension1: "value",
-  comexampleothervalue: 5,
+  id: expect.any(String),
+  time: expect.any(String),
   datacontenttype: "application/json",
   data: {
-    appinfoA: "abc",
-    appinfoB: 123,
-    appinfoC: true,
+    appinfoA: "abcd",
+    appinfoB: 1234,
+    appinfoC: false,
   },
-});
+};
+
+const cloudEvent2 = {
+  ...cloudEvent1,
+  specversion: "1.1",
+  type: "com.example.someevent.cloudEvent2",
+  source: "/mycontext",
+  subject: "example",
+  comexampleextension1: "value",
+  comexampleothervalue: 5,
+  datacontenttype: "application/cloudevents+json",
+};
+
+const cloudEvent3 = {
+  ...cloudEvent2,
+  type: "com.example.someevent.cloudEvent3",
+};
+
+const cloudEvent4 = {
+  ...cloudEvent3,
+  type: "com.example.someevent.cloudEvent4",
+};
+
+const cloudEvent5 = {
+  ...cloudEvent4,
+  type: "com.example.someevent.cloudEvent5",
+};
 
 describe("CloudEvent", () => {
   let socket;
 
   beforeAll(async () => {
     socket = await connect(
-      "/ws/cloudevent",
+      "cloudevent",
       {},
       {
         "Sec-WebSocket-Protocol": "cloudevents.json",
       },
-      ["cloudevents.json"],
     );
   });
 
@@ -98,24 +97,48 @@ describe("CloudEvent", () => {
   });
 
   test("Cloud Event Protocol", async () => {
-    expect(socket._protocol).toEqual("cloudevents.json");
+    expect(socket._opts.extraHeaders["Sec-WebSocket-Protocol"]).toEqual("cloudevents.json");
+    expect(socket.serverSocket.handshake.headers["sec-websocket-protocol"]).toEqual("cloudevents.json");
   });
 
-  test("Cloud event", async () => {
+  test("Cloud event (modeling)", async () => {
     const waitCloudEvent1Promise = waitForEvent(socket, "cloudEvent1");
     const waitCloudEvent2Promise = waitForEvent(socket, "cloudEvent2");
     const waitCloudEvent3Promise = waitForEvent(socket, "cloudEvent3");
     const waitCloudEvent4Promise = waitForEvent(socket, "cloudEvent4");
-    const result = await emitEvent(socket, "sendCloudEvent", cloudEvent1Message);
+    const waitCloudEvent5Promise = waitForEvent(socket, "cloudEvent5");
+    const result = await emitEvent(socket, "sendCloudEventModel", cloudEventModel);
     expect(result).toBe(true);
     const waitResult1 = await waitCloudEvent1Promise;
-    expect(waitResult1).toEqual(cloudEvent1Message);
+    expect(waitResult1).toEqual(cloudEvent1);
     const waitResult2 = await waitCloudEvent2Promise;
-    expect(waitResult2).toEqual(cloudEvent2Message);
+    expect(waitResult2).toEqual(cloudEvent2);
     const waitResult3 = await waitCloudEvent3Promise;
-    expect(waitResult3).toEqual(cloudEvent3Message);
+    expect(waitResult3).toEqual(cloudEvent3);
     const waitResult4 = await waitCloudEvent4Promise;
-    expect(waitResult4).toEqual(cloudEvent4Message);
+    expect(waitResult4).toEqual(cloudEvent4);
+    const waitResult5 = await waitCloudEvent5Promise;
+    expect(waitResult5).toEqual(cloudEvent5);
+  });
+
+  test("Cloud event (mapping)", async () => {
+    const waitCloudEvent1Promise = waitForEvent(socket, "cloudEvent1");
+    const waitCloudEvent2Promise = waitForEvent(socket, "cloudEvent2");
+    const waitCloudEvent3Promise = waitForEvent(socket, "cloudEvent3");
+    const waitCloudEvent4Promise = waitForEvent(socket, "cloudEvent4");
+    const waitCloudEvent5Promise = waitForEvent(socket, "cloudEvent5");
+    const result = await emitEvent(socket, "sendCloudEventMap", cloudEventMap);
+    expect(result).toBe(true);
+    const waitResult1 = await waitCloudEvent1Promise;
+    expect(waitResult1).toEqual(cloudEvent1);
+    const waitResult2 = await waitCloudEvent2Promise;
+    expect(waitResult2).toEqual(cloudEvent2);
+    const waitResult3 = await waitCloudEvent3Promise;
+    expect(waitResult3).toEqual({ ...cloudEvent3, subject: "cloud-example" });
+    const waitResult4 = await waitCloudEvent4Promise;
+    expect(waitResult4).toEqual({ ...cloudEvent4, subject: "cloud-example" });
+    const waitResult5 = await waitCloudEvent5Promise;
+    expect(waitResult5).toEqual({ ...cloudEvent5, subject: "cloud-example" });
   });
 
   test("Cloud event format error", async () => {